requirements.txt
# Python dependencies for HelpingHome project

indistinguishable_from_magic==0.0.2
opennote>=2.3.3
python-dotenv
elevenlabs
flask
flask-cors